+++ src/features/auth/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./login.ts";
+export * from "./signUp.ts";
